requirements.txt
discord.py[voice]>=2.4.0
yt-dlp
PyNaCl
audioop-lts
python-dotenv